initial migration to typescript
-added config files for vscode and extensions
-source code will now be in 'src' folder
-event handlers are working fine
