+++ .prettierrc
@@ -1,0 +1,20 @@
+{
+    "editor.formatOnSave": true,
+    "printWidth": 120,
+    "semi": true,
+    "singleQuote": true,
+    "tabWidth": 4,
+    "trailingComma": "es5",
+    "useTabs": false,
+    "jsxSingleQuote": false,
+    "endOfLine": "crlf",
+    "quoteProps": "as-needed",
+    "overrides": [
+        {
+            "files": "*.json",
+            "options": {
+                "tabWidth": 2
+            }
+        }
+    ]
+}
