+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "editor.tabSize": 4,
+  "editor.defaultFormatter": "esbenp.prettier-vscode"
+}
