--- commands/music/resume.js
@@ -1,31 +1,0 @@
-import { okEmoji } from '../../res/config.js';
-
-
-export default {
-
-    name: 'resume',
-    aliases: [],
-    description: 'resume the current song.',
-
-
-
-
-
-    async execute(Hina, msg, args) {
-
-        const profile = Hina.musicGuildProfile.get(msg.guildId);
-        if (!profile) return await msg.reply('I\'m not currently playing in this server!');
-
-        await profile.updateChannels(null, msg.channel);
-        await profile.player.unpause();
-        await msg.react(okEmoji);
-    },
-
-
-
-
-
-    async slashExecute(Hina, interaction) {
-        return;
-    },
-};
