+++ package.json
@@ -12,11 +12,12 @@
     "bufferutil": "^4.0.6",
     "detectlanguage": "^2.1.0",
     "discord.js": "^13.6.0",
+    "dotenv": "^16.0.0",
     "express": "^4.17.2",
     "ffmpeg-static": "^4.4.1",
+    "libsodium-wrappers": "^0.7.9",
     "node-fetch": "^3.2.0",
     "piston-client": "^1.0.2",
-    "sodium": "^3.0.2",
     "translate": "^1.4.1",
     "utf-8-validate": "^5.0.8",
     "yt-search": "^2.10.3",
@@ -25,11 +26,12 @@
   "devDependencies": {
     "@types/express": "^4.17.13",
     "@types/node": "^17.0.21",
+    "@types/yt-search": "^2.3.2",
     "node": "^16.13.2",
     "typescript": "^4.5.5"
   },
   "scripts": {
-    "start": "node .",
+    "start": "node ./dist/bot.js",
     "node-update": "npm i --save-dev node@16 && npm config set prefix=$(pwd)/node_modules/node && export PATH=$(pwd)/node_modules/node/bin:$PATH",
     "node-clean": "rm -rf node_modules && rm package-lock.json && npm cache clear --force && npm cache clean --force && npm i",
     "node-update-then-clean": "npm run node-update && npm run node-clean"
