--- res/config.js
@@ -1,6 +1,0 @@
-export const token = process.env.TOKEN;
-export const prefix = 'hina ';
-export const hinaColor = '#E49CFF';
-
-export const okEmoji = '902096184645124146';
-export const hinaImageOption = {dynamic: true, size: 4096};
