+++ types/hina/index.d.ts
@@ -1,0 +1,5 @@
+declare module 'hina' {
+    export interface CommandArgumentClassConstructor {
+        required?: Boolean;
+    }
+}
